refactor

+++ client/app/bundles/pages/UserProfilePage/components/AskList/index.jsx
@@ -4,10 +4,13 @@ import Modal from 'react-modal';
 
 import * as api from 'utilities/api';
 
-import AskListItem from '../AskListItem';
-import EditGift from 'components/EditGift/EditGift';
 import ConfirmationModal from 'common/modals/ConfirmationModal';
+import EditGift from 'common/components/EditGift';
+import LockedOverlay from 'common/components/LockedOverlay';
 
+import AskListItem from '../AskListItem';
+
+import * as layout from 'common/_styles/layout';
 import * as styled from './_styles';
 
 const AskList = (props) => {
@@ -15,21 +18,20 @@ const AskList = (props) => {
     editingElement.current.scrollIntoView();
   }
 
-  const [isModalOpen, setModalOpen] = React.useState(false);
+  const [scrollToItem, setScrollToItem] = React.useState(null);
   const [isAdding, setIsAdding] = React.useState(false);
-  const [editingItem, setEditingItem] = React.useState(null);
   const [deletingItem, setDeletingItem] = React.useState(null);
-  const [giftAction, setGiftAction] = React.useState(null);
   const [items, setItems] = React.useState({});
-  const [isApi, setIsApi] = React.useState(false);
+  const [isLocked, setIsLocked] = React.useState(false);
 
   const editingElement = useRef(null);
 
-  const editItemHandler = (e) => {
-    const parent = e.currentTarget.closest('[data-item-index]');
-    const item = items[parent.dataset.itemIndex];
-    setEditingItem(item);
-    openModal();
+  const editItemHandler = (newItem) => {
+    const newItems = items.map(item => {
+        return (item.id === newItem.id) ? newItem : item;
+    });
+    setItems(newItems);
+    setScrollToItem(newItem.id);
   }
 
   const deleteHandler = (e) => {
@@ -45,11 +47,17 @@ const AskList = (props) => {
   }
 
   const confirmDeleteHandler = () => {
+    if (isLocked) {
+      return;
+    }
+    setIsLocked(true);
+
     api.deleteGift({
       gift: deletingItem,
     }).then(response => {
       const newItems = items.filter(item => item.id !== deletingItem.id);
       setDeletingItem(null);
+      setIsLocked(false);
       setItems(newItems);
     });
   }
@@ -66,55 +74,65 @@ const AskList = (props) => {
       });
   }, []);
 
-
   useEffect(() => {
     if (isAdding) {
       scrollToEdit();
     }
   }, [isAdding]);
 
+  useEffect(() => {
+    if (scrollToItem) {
+      const el = document.querySelector(`[data-item-id="${scrollToItem}"]`);
+      if (el) {
+        el.scrollIntoView();
+      }
+    }
+  }, [scrollToItem]);
+
   return (
     <styled.Component>
-      <styled.TopContainer ref={ editingElement }>
-        { !isAdding &&
-          <styled.AddButton onClick={() => {
-            setIsAdding(true);
-          }}>Add to list</styled.AddButton>
-        }
-        
-        { isAdding &&
-          <EditGift
-            isNew={ true }
-            apiSave={ api.addAskingGift }
-            saveHandler={ addNewGiftHandler }
-            cancelHandler={ () => { setIsAdding(false); } }
-            />
-        }
-      </styled.TopContainer>
-
+      <layout.GridRow>
+        <styled.TopContainer ref={ editingElement }>
+          { !isAdding &&
+            <styled.AddButton onClick={() => {
+              setIsAdding(true);
+            }}>Add to list</styled.AddButton>
+          }
+          
+          { isAdding &&
+            <EditGift
+              apiSave={ api.addAskingGift }
+              saveHandler={ addNewGiftHandler }
+              cancelHandler={ () => { setIsAdding(false); } }
+              />
+          }
+        </styled.TopContainer>
+      </layout.GridRow>
       
       <styled.List>
         {items.length && items.map((item, index) => (
-          <styled.ListItem key={ item.id }>
+          <styled.ListItem key={ item.id } data-item-id={ item.id }>
             <AskListItem
               index={ index }
+              gift={ item }
               editHandler={ editItemHandler }
               deleteHandler={ deleteHandler }
-              {...item}
               />
           </styled.ListItem>
         ))}
       </styled.List>
-
-      <styled.BottomContainer>
-        <styled.AddButton onClick={() => {
-          if (isAdding) {
-            scrollToEdit();
-          } else {
-            setIsAdding(true);
-          }
-        }}>Add to list</styled.AddButton>
-      </styled.BottomContainer>
+      
+      <layout.GridRow>
+        <styled.BottomContainer>
+          <styled.AddButton onClick={() => {
+            if (isAdding) {
+              scrollToEdit();
+            } else {
+              setIsAdding(true);
+            }
+          }}>Add to list</styled.AddButton>
+        </styled.BottomContainer>
+      </layout.GridRow>
       
       { deletingItem &&
         <ConfirmationModal
@@ -123,6 +141,10 @@ const AskList = (props) => {
         >
           <p>Delete <span className="gift-title">{ deletingItem.title }</span>?</p>
         </ConfirmationModal>
+      }
+
+      { isLocked &&
+        <LockedOverlay />
       }
     </styled.Component>
   );
